refactored codebase structure
