--- app/(home)/layout.tsx
@@ -1,17 +1,0 @@
-import Navbar from "@/components/Navbar";
-import Footer from "@/components/Footer";
-
-
-export default function HomeLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (   
-    <>
-      <Navbar />
-        <main>{children}</main>
-      <Footer/>
-    </> 
-  );
-}
