+++ app/(static)/contact/page.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Mail, Phone, MapPin } from 'lucide-react';
 import { ExpandableSection } from '@/components/ExpandableSection';
 import Section from '@/components/SectionLayout';
-import { PageHero } from '@/app/(home)/components/HeroSection';
+import { PageHero } from '@/components/landing-page/HeroSection';
 
 
 const FAQS = [
