+++ app/(static)/layout.tsx
