+++ app/blog/[id]/page.tsx
@@ -1,5 +1,6 @@
+import { PageHero } from '@/components/landing-page/HeroSection';
+import { BackButton } from '@/components/BackButton';
 import { getAllPostIds, getPostData } from '@/lib/blog';
-import Link from 'next/link';
 
 export async function generateStaticParams() {
   const paths = getAllPostIds();
@@ -20,46 +21,48 @@ export default async function Post({ params }: { params: { id: string } }) {
   const post = await getPostData(id);
   
   return (
-    <div className="container mx-auto px-4 py-12">
-      <Link href="/blog" className="text-primary hover:underline mb-8 inline-block">
-        ← Back to all posts
-      </Link>
+    <><PageHero
+      title={post.title}
+      description={post.excerpt || ''}
+      backgroundImage={post.coverImage || ''} 
+      />
       
-      <article className="max-w-3xl mx-auto">
-        <header className="mb-8">
-          <h1 className="text-3xl md:text-4xl font-serif font-bold mb-4">{post.title}</h1>
-          <div className="flex flex-wrap gap-3 items-center text-gray-600">
-            <time className="text-sm">{new Date(post.date).toLocaleDateString()}</time>
-            {post.author && (
-              <>
-                <span className="text-sm">•</span>
-                <span className="text-sm">By {post.author}</span>
-              </>
-            )}
-            {post.category && (
-              <>
-                <span className="text-sm">•</span>
-                <span className="text-xs px-2 py-1 bg-primaryLight text-primary rounded-full">{post.category}</span>
-              </>
-            )}
-          </div>
-        </header>
-        
-        {post.coverImage && (
-          <div className="mb-8">
-            <img 
-              src={post.coverImage} 
-              alt={post.title} 
-              className="w-full h-auto rounded-lg"
-            />
-          </div>
-        )}
-        
-        <div 
-          className="prose prose-lg max-w-none font-serif leading-tight [&>*]:m-0" 
-          dangerouslySetInnerHTML={{ __html: post.contentHtml }} 
-        />
-      </article>
-    </div>
+      <div className="container relative mx-auto px-4 py-12">
+        <BackButton className="absolute left-10" />
+
+        <article className="max-w-3xl mx-auto">
+          <header className="mb-8">
+            {/* <h1 className="text-3xl md:text-4xl font-serif font-bold mb-4">{post.title}</h1> */}
+            <div className="flex flex-wrap gap-3 items-center text-gray-600">
+              <time className="text-sm">{new Date(post.date).toLocaleDateString()}</time>
+              {post.author && (
+                <>
+                  <span className="text-sm">•</span>
+                  <span className="text-sm">By {post.author}</span>
+                </>
+              )}
+              {post.category && (
+                <>
+                  <span className="text-sm">•</span>
+                  <span className="text-xs px-2 py-1 bg-primaryLight text-primary rounded-full">{post.category}</span>
+                </>
+              )}
+            </div>
+          </header>
+
+          {/* {post.coverImage && (
+            <div className="mb-8">
+              <img
+                src={post.coverImage}
+                alt={post.title}
+                className="w-full h-auto rounded-lg" />
+            </div>
+          )} */}
+
+          <div
+            className="prose prose-lg max-w-none font-serif leading-tight"
+            dangerouslySetInnerHTML={{ __html: post.contentHtml }} />
+        </article>
+      </div></>
   );
 }
