+++ app/layout.tsx
@@ -8,12 +8,6 @@ const crimsonPro = Crimson_Pro({
   weight: ["200", "300", "400", "500", "600", "700", "800", "900"],
 });
 
-// const poppins = Poppins({
-//   subsets: ["latin"],
-//   variable: "--font-poppins",
-//   weight: ["200", "300", "400", "500", "600", "700", "800", "900"],
-// });
-
 export const metadata: Metadata = {
   title: "AINAS",
   description: "African Initiative for Nature-based Solutions",
