+++ components/landing-page/HeroSection.tsx
