+++ components/landing-page/ImpactSection.tsx
