+++ components/landing-page/PartnersSection.tsx
