+++ components/landing-page/WhatWeDoSection.tsx
